Fix warnings (#1406)
* Log query id in executeQuery; Better type mismatch error; change format in report tool

* Better log query_id

* fix message

* Use c++11 thread_local instaed of gcc's __thread

* lock mutex before notifying waiting thread in sync insertion into distributed [#CLICKHOUSE-3379]

* Cmake: fix build without downloaded submodules (#1379)

* fix

* ZooKeeper: fixed stack smashing with tryGet()

The tryGet() operation creates a 1MB buffer on stack. This may or
may not work depending on the default stack size for threads,
whether the stack protector is enabled or not, recursion depth,
and the actual value size.

This is probably going to slow down some ZK operations, but I don't
see how else this could work reliably with the existing API.

* increased timeout for test_insertion_sync_fails_with_timeout

* Update CHANGELOG_RU.md

* Update ZooKeeper.cpp

* Fix warnings

* Fixes

* Dont strip debug info from  asan, tsan and other builds except releases

* Fix asan error causd by test 00144

* Fix empty log message (#CLICKHOUSE-3378)

+++ dbms/src/Common/tests/gtest_rw_lock_fifo.cpp.cpp
@@ -67,7 +67,7 @@ TEST(Common, RWLockFIFO_1)
             Stopwatch watch(CLOCK_MONOTONIC_COARSE);
 
             std::list<std::thread> threads;
-            for (int thread = 0; thread < pool_size; ++thread)
+            for (size_t thread = 0; thread < pool_size; ++thread)
                 threads.emplace_back([=] () { func(pool_size, round); });
 
             for (auto & thread : threads)
@@ -139,7 +139,7 @@ TEST(Common, RWLockFIFO_PerfTest_Readers)
             };
 
             std::list<std::thread> threads;
-            for (int thread = 0; thread < pool_size; ++thread)
+            for (size_t thread = 0; thread < pool_size; ++thread)
                 threads.emplace_back(func);
 
             for (auto & thread : threads)
